public_github_scrapper/views/users_views.py
from rest_framework.permissions import AllowAny
from rest_framework.response import Response
from rest_framework.views import APIView
from rest_framework import status

class UserAPI(APIView):
    permission_classes = (AllowAny,)

    def get(self, request, username):
        pass
