requirements/base.txt
django
urllib3
attrs
gunicorn
django-countries
django-nested-admin
django-grappelli
requests
django-cors-headers
djangorestframework
django-environ
